requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.2.0
billiard==3.3.0.23
celery==3.1.23
cffi==1.7.0
codicefiscale==0.9
cryptography==1.4
defusedxml==0.4.1
demjson==2.2.4
Django==1.8.5
django-admin-bootstrapped==2.5.6
django-allauth==0.24.1
django-appconf==1.0.2
django-bootstrap-form==3.2
django-celery==3.1.17
django-crispy-forms==1.6.0
django-geoposition==0.2.2
django-italian-utils==0.3
django-phonenumber-field==1.0.0
gevent==1.1.1
greenlet==0.4.10
idna==2.1
jsonpickle==0.9.3
kombu==3.0.35
ndg-httpsclient==0.4.1
oauthlib==1.0.3
paho-mqtt==1.1
phonenumbers==7.2.8
Pillow==3.2.0
psycopg2==2.6.1
pusher==1.3.0
pyasn1==0.1.9
pycparser==2.14
pynma==1.0
pyOpenSSL==16.0.0
pyserial==3.0.1
python3-openid==3.0.9
pytz==2016.3
redis==2.10.5
requests==2.8.1
requests-oauthlib==0.5.0
requirements==0.1
simplejson==3.8.2
six==1.10.0
tornado==4.3
urllib3==1.16
virtualenv==15.0.1
